LayoutTests/fast/parser/entity-end-textarea-tag.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<textarea id="mytextarea01">&lt;/textarea></textarea><br>
<textarea id="mytextarea02"><&#x2f;textarea></textarea><br>
<textarea id="mytextarea03"></&#x74;extarea></textarea><br>
<textarea id="mytextarea04"></t&#x65;xtarea></textarea><br>
<textarea id="mytextarea05"></te&#x78;tarea></textarea><br>
<textarea id="mytextarea06"></tex&#x74;area></textarea><br>
<textarea id="mytextarea07"></text&#x61;rea></textarea><br>
<textarea id="mytextarea08"></texta&#x72;ea></textarea><br>
<textarea id="mytextarea09"></textar&#x65;a></textarea><br>
<textarea id="mytextarea10"></textare&#x61;></textarea><br>
<textarea id="mytextarea11"></textarea&gt;</textarea><br>
<p id="description"></p>
<div id="console"></div>
<script>
description("Test parsing of entity-escaped &lt;/textarea&gt; tag for <a href=\"https://bugs.webkit.org/show_bug.cgi?id=7931\">Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely</a>");

var expectedResult = '<' + '/textarea>';
shouldBe('document.getElementById("mytextarea01").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea02").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea03").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea04").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea05").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea06").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea07").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea08").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea09").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea10").textContent', 'expectedResult');
shouldBe('document.getElementById("mytextarea11").textContent', 'expectedResult');
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
