third_party/blink/web_tests/platform/generic/fast/parser/entity-end-textarea-tag-expected.txt














Test parsing of entity-escaped </textarea> tag for Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("mytextarea01").textContent is expectedResult
PASS document.getElementById("mytextarea02").textContent is expectedResult
PASS document.getElementById("mytextarea03").textContent is expectedResult
PASS document.getElementById("mytextarea04").textContent is expectedResult
PASS document.getElementById("mytextarea05").textContent is expectedResult
PASS document.getElementById("mytextarea06").textContent is expectedResult
PASS document.getElementById("mytextarea07").textContent is expectedResult
PASS document.getElementById("mytextarea08").textContent is expectedResult
PASS document.getElementById("mytextarea09").textContent is expectedResult
PASS document.getElementById("mytextarea10").textContent is expectedResult
PASS document.getElementById("mytextarea11").textContent is expectedResult
PASS successfullyParsed is true

TEST COMPLETE
